split up recipes.ts into different functions + added method headers
